my_project/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
cffi==1.16.0; python_version >= '3.8'
cryptography==41.0.7; python_version >= '3.7'
django==4.2.9; python_version >= '3.8'
django-cors-headers==4.3.1; python_version >= '3.8'
djangorestframework==3.14.0; python_version >= '3.6'
pycparser==2.21
pymysql==1.1.0; python_version >= '3.7'
pytz==2023.3.post1
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.9.0; python_version < '3.11'
